% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa00_squarebrackets_help.R
\name{aaa00_squarebrackets_help}
\alias{aaa00_squarebrackets_help}
\alias{squarebrackets-package}
\alias{squarebrackets}
\alias{squarebrackets_help}
\title{squarebrackets: Subset Methods as Alternatives to the Square Brackets Operators for Programming}
\description{
squarebrackets: \cr
Subset Methods as Alternatives to the Square Brackets Operators for Programming. \cr
\cr
Provides subset methods
(supporting both atomic and recursive S3 classes)
that may be more convenient alternatives to the \code{[} and \verb{[<-} operators,  whilst maintaining similar performance.

Some nice properties of these methods include, but are not limited to, the following:
\enumerate{
\item The \code{[} and \verb{[<-} operators use different rule-sets for different data.frame-like types
(data.frames, data.tables, tibbles, tidytables, etc.); \cr    The 'squarebrackets' methods use the same rule-sets for the different data.frame-like types.
\item Performing dimensional subset operations on an array using \code{[} and \verb{[<-},    requires a-priori knowledge on the number of dimensions the array has; \cr    The 'squarebrackets' methods work on any arbitrary dimensions without requiring such prior knowledge.
\item Unlike the \code{[} and \verb{[<-} operators,    the 'squarebrackets' methods will operate on duplicate names
(not just the first match),    use consistent syntax for inverting indices,    give an error for out-of-bounds indices,    and support the use of advanced indices through formulas.
\item The \verb{[<-} operator only supports copy-on-modify semantics for most classes; \cr    The 'squarebrackets' methods provide explicit pass-by-reference and pass-by-value semantics,    and do so safely.
\item 'squarebrackets' supports index-less sub-set operations for \verb{long vectors},    which is more memory efficient than sub-set operations using the \code{[} and \verb{[<-} operators.
\item All of the methods provided by 'squarbrackets' are programmatically friendly
(no Non-Standard Evaluation).\cr \cr
}
}
\section{Goal}{


Among programming languages,
'R' has perhaps one of the most
flexible and comprehensive sub-setting functionality,
provided by the square brackets operators (\code{[}, \verb{[<-}). \cr
But in some situations the square brackets operators
are occasionally less than optimally convenient \cr
\cr
The Goal of the 'squarebrackets' package
is not to replace the square-brackets operators,
but to provide \bold{alternative} sub-setting methods and functions,
to be used in situations where the square bracket operators are inconvenient. \cr
\cr
}

\section{Quick Start Guide}{

For the Quick Start Guide, see: \cr
\url{https://tony-aw.github.io/squarebrackets/articles/squarebrackets.html}. \cr
\cr
}

\section{Overview Help Pages}{


\bold{Essentials} \cr
The essential documentation is split into the following help pages:
\itemize{
\item \link{squarebrackets_methods}: \cr
Lists the main methods provided by 'squarebrackets'. \cr
Also explains the method dispatch system in 'squarebrackets'.
\item \link{squarebrackets_index_fundamentals}: \cr
Explains the essential fundamentals of the indexing forms in 'squarebrackets'.
\item \link{squarebrackets_keywords}: \cr
Explains the usage of keywords in the main methods of 'squarebrackets'. \cr \cr
}

\bold{Arguments} \cr
The methods in 'squarebrackets' share a lot of common arguments. \cr
The explanations for these common arguments are given in the following help pages:
\itemize{
\item \link{squarebrackets_supported_structures}: \cr
Lists the structures that are supported by 'squarebrackets',
and explains some related terminology.
\item \link{squarebrackets_index_args}: \cr
Explains the common indexing arguments used in the main S3 methods.
\item \link{squarebrackets_modify}: \cr
Explains the modification-related arguments,
and other essential information regarding modification.
\item \link{squarebrackets_options}: \cr
Lists and explains the options the user can specify in 'squarebrackets'.
\item \link{squarebrackets_stride}: \cr
Gives an overview of the \code{stride} argument in the \link[=long_x]{long_} methods. \cr \cr
}

\bold{Pass-By-Reference} \cr
The following help pages explain the pass-by-reference semantics provided by 'squarebrackets',
and only need to be read when planning to use those semantics:
\itemize{
\item \link{squarebrackets_PassByReference}: \cr
Explains Pass-by-Reference semantics, and its important consequences.
\item \link{squarebrackets_coercion}: \cr
Explains the difference in coercion rules between
modification through Pass-by-Reference semantics and
modification through copy (i.e. pass-by-value). \cr \cr
}
}

\section{Helper Functions}{


A couple of convenience functions, and helper functions for creating ranges, sequences, and indices
(often needed in sub-setting)
are provided:
\itemize{
\item \link{n}: Nested version of \link[base]{c},
and short-hand for \link[base]{list}.
\item \link{ndim}: Get the number of dimensions of an object.
\item \link{ss2coord}, \link{coord2ii}: Convert subscripts
(dimensional array indices) to coordinates,
coordinates to flat indices,
and vice-versa.
\item \link{match_all}: Find all matches, of one vector in another,
taking into account the order and any duplicate values of both vectors.
\item Computing indices: \cr
\link{idx_by} to compute grouped indices. \cr \cr
}
}

\section{Properties Details}{

The alternative sub-setting methods and functions provided by 'squarebrackets'
have the following properties:
\itemize{
\item \bold{Programmatically friendly}:
\itemize{
\item Unlike base \code{[},
it's not required to know the number of dimensions of an array a-priori,
to perform subset-operations on an array.
\item Missing arguments can be filled with \code{NULL},
instead of using dark magic like \code{base::quote(expr =    )}.
\item No Non-standard evaluation.
\item Functions are pipe-friendly.
\item No (silent) vector recycling.
\item Extracting and removing subsets uses the same syntax.
}
\item \bold{Class consistent}:
\itemize{
\item sub-setting of multi-dimensional objects by specifying dimensions
(i.e. rows, columns, ...)
use \code{drop = FALSE}. \cr
So matrix in, matrix out.
\item The methods deliver the same results for
data.frames, data.tables, tibbles, and tidytables. \cr
No longer does one have to re-learn the different brackets-based sub-setting rules
for different types of data.frame-like objects. \cr
Powered by the subclass agnostic 'C'-code from 'collapse' and 'data.table'.
}
\item \bold{Explicit copy semantics}:
\itemize{
\item Sub-set operations that change its memory allocations,
always return a modified (partial) copy of the object. \cr
\item For sub-set operations that just change values in-place
(similar to the \verb{[<-} and \verb{[[<-} methods)
the user can choose a method that modifies the object by \bold{reference},
or choose a method that returns a \bold{(partial) copy}.
}
\item \bold{Careful handling of names}:
\itemize{
\item Sub-setting an object by index names returns ALL matches with the given names,
not just the first.
\item Data.frame-like objects (see supported classes below)
are forced to have unique column names.
}
\item \bold{Concise function and argument names}.
\item \bold{Performance & Energy aware}: \cr
Despite the many checks performed, the functions are kept reasonably speedy,
through the use of the 'Rcpp', 'collapse', and 'data.table' R-packages. \cr
The functions were also made to be as memory efficient as reasonably possible,
to lower the carbon footprint of this package. \cr \cr
}
}

\references{
The badges shown in the documentation of this R-package were made using the services of: \url{https://shields.io/}
}
\author{
\strong{Author, Maintainer}: Tony Wilkes \email{tony_a_wilkes@outlook.com} (\href{https://orcid.org/0000-0001-9498-8379}{ORCID})
}
\concept{intro}
