% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa07_squarebrackets_ellipsis.R
\name{aaa07_squarebrackets_ellipsis}
\alias{aaa07_squarebrackets_ellipsis}
\alias{squarebrackets_ellipsis}
\title{Ellipses Usage in 'squarebrackets'}
\description{
Due to how the S3 method dispatch system works in 'R',
all generic methods have the ellipsis argument (\code{...}). \cr
For the user's safety,
'squarebrackets' does check that the user doesn't accidentally
add arguments that make no sense for that method. \cr
\cr
\cr
}
\concept{ellipsis}
