% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa09_squarebrackets_PassByReference.R
\name{aaa09_squarebrackets_PassByReference}
\alias{aaa09_squarebrackets_PassByReference}
\alias{squarebrackets_PassByReference}
\title{Regarding Modification By Reference}
\description{
This help page describes how modification using "pass-by-reference" semantics
is handled by the 'squarebrackets' package. \cr
\cr
"Pass-by-reference" refers to modifying a mutable object,
or a subset of a mutable object,
without making any copies at all. \cr
\cr
This help page does not explain all the basics of pass-by-reference semantics,
as this is treated as prior knowledge. \cr
All functions/methods in the 'squarebrackets' package
with the word "set" in the name
use pass-by-reference semantics. \cr \cr
}
\section{Advantages and Disadvantages}{

The main advantage of pass-by-reference is that much less memory is required to modify objects,
and modification is also generally faster. \cr
But it does have several disadvantages. \cr
\cr
First, the coercion rules are slightly different: see \link{squarebrackets_coercion}. \cr
\cr
Second, if 2 or more variables refer to exactly the same object
(i.e. have the same address),
changing one variable also changes the other ones. \cr
I.e. the following code,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- y <- mutatomic(1:16)
ii_set(x, 1:6, rp = 8)
}\if{html}{\out{</div>}}

modifies not just \code{x}, but also \code{y}. \cr
This is true even if one of the variables is locked
(see \link[base]{bindingIsLocked}). \cr
I.e. the following code,

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- mutatomic(1:16)
y <- x
lockBinding("y", environment())
ii_set(x, i = 1:6, rp = 8)
}\if{html}{\out{</div>}}

modifies both \code{x} and \code{y} without error,
even though \code{y} is a locked constant. \cr
\cr
}

\section{Mutable vs Immutable Classes}{

With the exception of environments,
most of base R's S3 classes are treated as immutable: \cr
Modifying an object in 'R' will make a copy of the object,
something called 'copy-on-modify' semantics. \cr
\cr
A prominent mutable S3 class is the \code{data.table} class,
which is a mutable data.frame class, and supported by 'squarebrackets'. \cr
Similarly, 'squarebrackets' adds a class for mutable atomic objects: \cr
\link{mutatomic}. \cr \cr
}

\section{Material vs Immaterial objects}{

Most objects in 'R' are material objects: \cr
the values an object contains are actually stored in memory. \cr
For example, given \code{x <- rnorm(1e6)}, \code{x} is a material object: \cr
1 million values (decimal numbers, in this case) are actually stored in memory. \cr
\cr
In contrast, \link[=makeActiveBinding]{ActiveBindings} are immaterial: \cr
They are objects that,
when accessed,
call a function to generate values on the fly,
rather than actually storing values. \cr
\cr
Since immaterial objects do not actually store the values in memory,
the values obviously also cannot be changed in memory. \cr
Therefore, Pass-by-Reference semantics don't work on immaterial objects. \cr
\cr
}

\section{ALTREP}{

The \link{mutatomic} constructors
(i.e. \link{mutatomic}, \link{as.mutatomic}, etc.)
will automatically materialize ALTREP objects,
to ensure consistent behaviour for 'pass-by-reference' semantics. \cr
\cr
A \code{data.table} can have ALTREP columns. \cr
A \code{data.tables} will coerce the column to a materialized column when it is modified, even by reference. \cr
\cr
}

\section{Input Variable}{

Methods/functions that perform in-place modification by reference
only works on objects that actually exist as an actual variable,
similar to functions in the style of \code{some_function(x, ...) <- value}. \cr
Thus things like any of the following, \cr
\code{ii_set(1:10, ...)}, \code{ii_set(x$a, ...)}, or \code{ii_set(base::letters)}, \cr
will not and should not work. \cr \cr
}

\section{Lock Binding}{

Mutable classes are,
as the name suggests,
meant to be mutable. \cr
Locking the binding of a mutable object is fruitless. \cr
To ensure an object cannot be modified by any of the methods/functions from 'squarebrackets',
2 things must be true: \cr
\itemize{
\item the object must be an immutable class.
\item the binding must be \bold{locked} (see \link[base]{lockBinding}). \cr \cr
}
}

\section{Protection}{


Due to the properties described above in this help page,
'squarebrackets' protects the user from do something like the following:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
# letters = base::letters
ii_set(letters, i = 1, rp = "XXX")

}\if{html}{\out{</div>}}

'squarebrackets' will give an error when running the code above, because:
\enumerate{
\item most addresses in \code{baseenv()} are protected;
\item immutable objects are disallowed
(you'll have to create a mutable object,
which will create a copy of the original,
thus keeping the original object safe from modification by reference);
\item locked bindings are disallowed.
}
}

\examples{

# the following code demonstrates how locked bindings,
# such as `base::letters`,
# are being safe-guarded

x <- list(a = base::letters)
myref <- x$a # view of a list
address(myref) == address(base::letters) # TRUE: point to the same memory
bindingIsLocked("letters", baseenv()) # base::letters is locked ...
bindingIsLocked("myref", environment()) # ... but this pointer is not!

if(requireNamespace("tinytest")) {
  tinytest::expect_error(
    ii_set(myref, i = 1, rp = "XXX") # this still gives an error though ...
  )
}

is.mutatomic(myref) # ... because it's not of class `mutatomic`


x <- list(
  a = as.mutatomic(base::letters) # `as.mutatomic()` makes a copy
)
myref <- x$a # view of a list
address(myref) == address(base::letters) # FALSE: it's a copy
ii_set(
  myref, i = 1, rp = "XXX"  # modifies x, does NOT modify `base::letters`
)
print(x) # x is modified
base::letters # but this still the same


}
