% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa4_squarebrackets_duplicates.R
\name{aaa4_squarebrackets_duplicates}
\alias{aaa4_squarebrackets_duplicates}
\alias{squarebrackets_duplicates}
\title{On Duplicates}
\description{
The \link{sb_x} method is the only method where providing duplicate indices actually make sense. \cr
For the other methods, it doesn't make sense. \cr
Giving duplicate indices usually won't break anything;
however, when replacing/transforming or removing subsets,
it is almost certainly not the intention to provide duplicate indices. \cr
Providing duplicate indices anyway might lead to unexpected results. \cr
Therefore, for the methods where giving duplicate indices does not make sense,
the \code{chkdup} argument is present. \cr
This argument controls whether the method in question checks for duplicates (\code{TRUE}) or not (\code{FALSE}). \cr
\cr
Setting \code{chkdup = TRUE} means the method in question will check for duplicate indices,
and give an error when it finds them. \cr
\cr
Setting \code{chkdup = FALSE} will disable these checks,
which saves time and computation power, and is thus more efficient. \cr
\cr
Since checking for duplicates can be expensive, it is set to \code{FALSE} by default. \cr \cr
}
