% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayshadow.R
\name{ArrayShadow}
\alias{ArrayShadow}
\alias{cast_ArrayShadow}
\alias{cast_ArrayShadow2}
\title{Array Shadow}
\usage{
cast_ArrayShadow(type, length, dim, names, dimnames, oldClass)

cast_ArrayShadow2(x_expr, env)
}
\arguments{
\item{type}{the type of the vector/array (see \link[base]{atomic} and \link[base]{list}).}

\item{length}{the length of the vector/array.}

\item{dim}{the dimensions of the vector/array.}

\item{names}{the names of the vector/array.}

\item{dimnames}{the \code{dimnames} of the vector/array.}

\item{oldClass}{the class attribute of the vector/array.}

\item{x_expr}{the substituted name of the vector/array.}

\item{env}{the environment where the vector/array name (\code{x_expr}) actually lives. \cr}
}
\value{
The "shadow" of the vector/array.
}
\description{
The Array Shadow object is a "shadow" of an vector or array \code{x}: \cr
It has the same (dim)names as \code{x}, and pretends to have the same length and dimensions. \cr
But it holds no actual data. \cr
}
\examples{
x <- array(1:27, c(3,3,3))
cast_ArrayShadow(typeof(x), length(x), dim(x), names(x), dimnames(x), oldClass(x))
}
