% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_seq.R
\name{ci_seq}
\alias{ci_seq}
\title{Construct Sequence Parameters Based on Margins}
\usage{
ci_seq(x, m = 0L, start = NULL, end = NULL, by = 1L, .abortcall = sys.call())
}
\arguments{
\item{x, m, start, end, by}{see \link{idx_r}.}

\item{.abortcall}{environment where the error message is passed to. \cr \cr}
}
\value{
A list of sequence parameters. \cr \cr
}
\description{
\code{ci_seq()} returns a list of parameters to construct a sequence based on the margins of an object. \cr
It is internally used by the \link{idx_r} function. \cr \cr
}
\examples{

x <- data.frame(
  a = 1:10, b = letters[1:10], c = factor(letters[1:10]), d = -1:-10
)
print(x)
ind1 <- idx_r(x, 1, 2, 2-1i) # rows 2:(nrow(x)-1)
sb2_x(x, ind1) # extract the row range

x <- array(1:125, c(5,5,5))
dims <- 1:3
sub <- idx_r(x, dims, 2, 2-1i) # 2:(n-1) for every dimension
sb_x(x, sub, dims) # same as x[ 2:4, 2:4, 2:4, drop = FALSE]

x <- letters
x[idx_r(x, 0, 2, 2-1i)]

}
