% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_recursive.R
\name{dropl}
\alias{dropl}
\title{Helper Functions for Sub-Set Operations on Recursive Objects}
\usage{
dropl(x)
}
\arguments{
\item{x}{a list.}
}
\value{
For \code{dropl()}: \cr
If \code{length(x) == 1L}, \code{dropl(x)} returns \code{x[[1L]]}; \cr
otherwise it returns the original \code{x} unchanged. \cr
\cr
}
\description{
'squarebrackets' provides some helper functions for sub-set operations on recursive objects
(lists and data.frames). \cr
\cr
\code{dropl(x)} returns \code{x[[1L]]} if \code{length(x) == 1}, and returns \code{x} otherwise. \cr
This can be used for the \verb{*_x} and \verb{*_wo} methods on recursive objects,
where one would like to extract the contents of the singular selection. \cr
\cr
}
\examples{

obj <- as.list(1:10)
ii_x(obj, 1) |> dropl() # equivalent to obj[[1L]]


}
