% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_by.R
\name{idx_by}
\alias{idx_by}
\title{Compute Grouped Indices}
\usage{
idx_by(x, m, f, grp, parallel = FALSE, mc.cores = 1L)
}
\arguments{
\item{x}{the object from which to compute the indices.}

\item{m}{a single non-negative integer giving the margin for which to compute indices. \cr
For flat indices or for non-dimensional objects, use \code{m = 0L}. \cr}

\item{f}{a subset function to be applied per group on \code{indices}. \cr
If \code{m == 0L}, \code{indices} is here defined as \code{setNames(1:length(x), names(x))}. \cr
If \code{m > 0L}, \code{indices} is here defined as \code{setNames(1:dim(x)[m], dimnames(x)[[m]])}. \cr
The function must produce a character or integer vector as output. \cr
For example, to subset the last element per group, specify: \cr
\code{f = last}}

\item{grp}{a factor giving the groups.}

\item{parallel, mc.cores}{see \link[collapse]{BY}.}
}
\value{
A vector of indices.
}
\description{
Given:
\itemize{
\item a sub-set function \code{f};
\item an object \code{x} with its margin \code{m};
\item and a grouping factor \code{grp};
}

the \code{idx_by()} function takes \code{indices} \bold{per group} \code{grp}. \cr
The result of \code{idx_by()} can be supplied to the indexing arguments
(see \link{squarebrackets_index_args})
to perform \bold{grouped} subset operations. \cr
\cr
}
\examples{


# vectors ====
(a <- 1:20)
(grp <- factor(rep(letters[1:5], each = 4)))

# get the last element of `a` for each group in `grp`:
s <- list(idx_by(a, 0L, last, grp))
ss_x(cbind(a, grp), s, 1L)


# data.frame ====
x <- data.frame(
  a = sample(1:20),
  b = letters[1:20],
  group = factor(rep(letters[1:5], each = 4))
)
print(x)
# get the first row for each group in data.frame `x`:
row <- idx_by(x, 1, first, x$group)
tt_x(x, row)
# get the first row for each group for which a > 10:
x2 <- tt_x(x, with(x,  a > 10))
row <- na.omit(idx_by(x2, 1, first, x2$group))
tt_x(x2, row)
 

}
