% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_dt.R
\name{idx_obs}
\alias{idx_obs}
\alias{idx_vars}
\title{Compute Special Indices for data.frames}
\usage{
idx_obs(x, form, use = 1)

idx_vars(x, form, use = 1)
}
\arguments{
\item{x}{a data.frame}

\item{form}{a formula; see \code{Examples} section below.}

\item{use}{set to \code{1} to use the specified indices,
set to \code{-1} to use all indices \strong{except} the specified indices. \cr \cr}
}
\value{
A vector of indices.
}
\description{
\code{idx_obs()} and \code{idx_vars()} produce an integer vector of row or column indices, respectively,
given formulas based on data.frame variables. \cr
\cr
}
\examples{


obj <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
print(obj)
sbt_x(obj, 1:3, 1:3)
sbt_x(obj, idx_obs(obj, ~ (a > 5) & (c < 19)), idx_vars(obj, a ~ c))


}
