% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_r.R
\name{idx_r}
\alias{idx_r}
\title{Compute Heterogeneous Index Range}
\usage{
idx_r(x, m, start, end, by = 1L)
}
\arguments{
\item{x}{the object from which to compute the indices.}

\item{m}{a single non-negative integer giving the margin for which to compute indices. \cr
For flat indices or for non-dimensional objects, use \code{m = 0L}. \cr}

\item{start, end}{the start and end of the range. \cr
Can be any combination of any of the following: \cr
\itemize{
\item A strictly positive integer.
\item A complex number (see explanation in \link{squarebrackets_indx_args}).
\item A string to refer to a name. \cr
In case of \code{start}, the first  match will be used. \cr
In case of \code{end}, the last match will be used. \cr
\item a Boolean-returning function to be applied on \code{indices}. \cr
If \code{m == 0L}, \code{indices} is here defined as \code{setNames(1:length(x), names(x))}. \cr
If \code{m > 0L}, \code{indices} is here defined as \code{setNames(1:dim(x)[m], dimnames(x)[[m]])}. \cr
For \code{start}, the first \code{TRUE} match will be used. \cr
For \code{end}, the last \code{TRUE} match will be used. \cr
}}

\item{by}{an optional single integer, giving the step size. \cr
\code{idx_r()} will automatically make sure the sign of \code{by} is set correctly.}
}
\value{
A vector of integer indices.
}
\description{
The \code{idx_r()} function computes an index range,
where one can combine different types for the start and end point. \cr \cr
}
\examples{

x <- data.frame(
  a = 1:10, b = letters[1:10], c = factor(letters[1:10]), d = -1:-10
)
print(x)
ind1 <- idx_r(x, 1, 2, 2-1i) # rows 2:(n-1)
ind2 <- idx_r(x, 2, "d", 2) # columns d:2
sb2_x(x, ind1, ind2)
}
