% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_rng.R
\name{idx_rng}
\alias{idx_rng}
\title{Compute Heterogeneous Index Range}
\usage{
idx_rng(x, start, end, margin = 0L, by = 1L)
}
\arguments{
\item{x}{the object from which to compute the indices.}

\item{start, end}{the start and end of the range. \cr
Can be any combination of any of the following: \cr
\itemize{
\item A strictly positive integer.
\item A complex number (see explanation in \link{squarebrackets_indx_args}).
\item A string to refer to a name. \cr
In case of \code{start}, the first  match will be used. \cr
In case of \code{end}, the last match will be used. \cr
\item a Boolean-returning function to be applied on \code{setNames(indices, names(indices))}. \cr
In case of \code{start}, the first \code{TRUE} match will be used. \cr
In case of \code{end}, the last \code{TRUE} match will be used. \cr
}}

\item{margin}{giving the margin for which to compute indices. \cr
For flat indices or for non-dimensional objects, keep \code{margin = 0}. \cr}

\item{by}{an optional single integer, giving the step size.}
}
\value{
A vector of integer indices.
}
\description{
The \code{idx_rng()} function computes an index range,
where one can combine different types for the start and end point. \cr \cr
}
\examples{

x <- data.frame(
  a = 1:10, b = letters[1:10], c = factor(letters[1:10]), d = -1:-10
)
print(x)
ind1 <- idx_rng(x, 2, 2-1i, 1) # rows 2:(n-1)
ind2 <- idx_rng(x, "d", 2, 2) # columns d:2
sb2_x(x, ind1, ind2)
}
