% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.formula.R
\name{is.formula}
\alias{is.formula}
\title{Check if an Object is a Formula}
\usage{
is.formula(form)
}
\arguments{
\item{form}{object to check}
}
\value{
The \code{is_formula()} function returns \code{TRUE} if the input is a formula,
and \code{FALSE} otherwise. \cr \cr
}
\description{
\code{is.formula()} checks if an object is a formula. \cr \cr
}
\examples{
is.formula(~ x)
is.formula(1:10)

}
