% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\name{keywords}
\alias{keywords}
\alias{is.formula}
\title{Formula with Keywords}
\usage{
is.formula(form)
}
\value{
The \code{is_formula()} function returns \code{TRUE} if the input is a formula,
and \code{FALSE} otherwise. \cr \cr
}
\description{
Instead of the usual indexing types,
formulas can be used to specify indices in arrays that can evaluate keywords. \cr
\cr
The following keywords are available:
\itemize{
\item \code{.M}: the given margin/dimension; \code{0} if not relevant.
\item \code{.Nms}: the (dim)names at the given margin.
\item \code{.N}: the size of a given dimension (if \code{.M} is not \code{0}) or else the length of \code{x}.
\item \code{.I}: equal to \code{1:.N}.
\item \code{.bi(...)}: a \bold{function} to specify bilateral indices.
\item \code{.x}: the input variable \code{x} itself.
Here one can specify both positive and negative numbers. \cr
Positive numbers specify indices work like normally. \cr
Negative numbers specify indices from the end. \cr
}

These keywords can be used inside formulas to specify more advanced indices. \cr

Here are a few examples of advanced indexing in arrays:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- matrix(1:20, 5, 4)
dimnames(x) <- list(month.abb[1:5], month.abb[1:4])
print(x)
#>     Jan Feb Mar Apr
#> Jan   1   6  11  16
#> Feb   2   7  12  17
#> Mar   3   8  13  18
#> Apr   4   9  14  19
#> May   5  10  15  20


ss_x(x, ~ 1:round(.N/2)) # select first half of rows and columns
#>     Jan Feb
#> Jan   1   6
#> Feb   2   7

ss_x(x, ~ grep("a", .Nms)) # select subset whose dimnames contains an "a"
#>     Jan Mar
#> Jan   1  11
#> Mar   3  13
#> May   5  15

ss_x(x, ~ .bi(1, -1), -Inf) # remove first and last index of every dimension 
#>     Feb Mar
#> Feb   7  12
#> Mar   8  13
#> Apr   9  14
}\if{html}{\out{</div>}}
}
\examples{
x <- matrix(1:20, 5, 4)
dimnames(x) <- list(month.abb[1:5], month.abb[1:4])
print(x)


ss_x(x, ~ 1:round(.N/2)) # select first half of rows and columns

ss_x(x, ~ grep("a", .Nms)) # select subset whose dimnames contains an "a"

ss_x(x, ~ .bi(c(1, -1)), -Inf) # remove first and last index of every dimension 



}
