% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lst_rec.R
\name{lst_rec}
\alias{lst_rec}
\alias{lst_rec.default}
\alias{lst_recin}
\alias{lst_recin.default}
\title{Access, Replace, Transform, Delete, or Extend Recursive Subsets}
\usage{
lst_rec(x, ...)

\method{lst_rec}{default}(x, rec, ...)

lst_recin(x, ...)

\method{lst_recin}{default}(x, rec, ..., rp, tf)
}
\arguments{
\item{x}{a list, or list-like object.}

\item{...}{see \link{squarebrackets_ellipsis}.}

\item{rec}{a strictly positive integer vector or character vector, of length \code{p},
such that \code{lst_rec(x, rec)} is equivalent to \verb{x[[ rec[1] ]]...[[ rec[p] ]]},
providing all but the final indexing results in a list. \cr
When on a certain subset level of a nested list,
multiple subsets with the same name exist,
only the first one will be selected when performing recursive indexing by name,
since recursive indexing can only select a single element. \cr
\verb{NA, NaN, Inf, -Inf, NULL} are not valid values for \code{rec}.}

\item{rp}{optional, and allows for multiple functionalities:
\itemize{
\item In the simplest case, performs \code{x[[rec]] <- rp},
using R's native semantics. \cr
Since this is a replacement of a recursive subset,
\code{rp} does not necessarily have to be a list itself; \cr
\code{rp} can be any type of object.
\item Specifying \code{rp = NULL} will \bold{delete} (recursive) subset \code{lst_rec(x, rec)}. \cr
To specify actual \code{NULL} instead of deleting a subset, use \code{rp = list(NULL)}.
\item When \code{rec} is an integer, and specifies an out-of-bounds subset,
\code{lst_recin()} will add value \code{rp} to the list. \cr
Any empty positions in between will be filled with \code{NA}.
\item When \code{rec} is character, and specifies a non-existing name,
\code{lst_recin()} will add value \code{rp} to the list as a new element at the end.
}}

\item{tf}{an optional function. If specified, performs \code{x[[rec]] <- tf(x[[rec]])},
using R's native semantics. \cr
Does not support extending a list like argument \code{rp}.}
}
\value{
For \code{lst_rec()}: \cr
Returns the recursive subset. \cr
\cr
For \code{lst_recin(..., rp = rp)}: \cr
Returns nothing,
but replaces, adds, or deletes the specified recursive subset,
using R's native semantics. \cr
\cr
For \code{lst_recin(..., tf = tf)}: \cr
Returns nothin,
but transforms the specified recursive subset,
using R's native semantics. \cr \cr
}
\description{
The \code{lst_rec()} and \code{lst_recin()} methods
are essentially convenient wrappers around \code{[[} and \verb{[[<-},
respectively. \cr
\cr
\code{lst_rec()} will access recursive subsets of lists. \cr
\cr
\code{lst_recin()} can do the following things: \cr
\itemize{
\item replace or transform recursive subsets of a list,
using R's native semantics,
by specifying the \code{rp} or \code{tf} argument, respectively.
\item delete a recursive subset of a list,
using R's native semantics,
by specifying argument \code{rp = NULL}.
\item extending a list with additional recursive elements,
using R's native semantics. \cr
This is done by specifying an out-of-bounds index in argument \code{rec},
and entering the new values in argument \code{rp}. \cr
Note that adding surface level elements of a dimensional list
will delete the dimension attributes of that list. \cr \cr
}
}
\details{
Since recursive objects are references to other objects,
extending a list or deleting an element of a list does not copy the entire list,
in contrast to atomic vectors. \cr
\cr
\bold{Tip}: \cr
Dimensional sub-set operations on dimensional lists
are much faster and more flexible
than Recursive sub-set operations on nested lists. \cr
So, whenever it makes sense, consider turning your nested list into a dimensional list. \cr
One can turn a hierarchical list into a dimensional list using,
for example,
the \href{https://tony-aw.github.io/broadcast/man/cast_hier2dim.html}{broadcast::cast_hier2dim} method from the 'broadcast' 'R' package. \cr
\cr
}
\examples{


lst <- list(
  A = list(
    A = list(A = "AAA", B = "AAB"),
    A = list(A  = "AA2A", B = "AA2B"),
    B = list(A = "ABA", B = "ABB")
  ),
  B = list(
    A = list(A = "BAA", B = "BAB"),
    B = list(A = "BBA", B = "BBB")
  ),
  C = list(
    A = 1:10,
    B = 11:20
  )
)

#############################################################################

# access recursive subsets ====

lst_rec(lst, c(1,2,2)) # this gives "AA2B"
lst_rec(lst, c("A", "B", "B")) # this gives "ABB"
lst_rec(lst, c(2,2,1)) # this gives "BBA"
lst_rec(lst, c("B", "B", "A")) # this gives "BBA"


#############################################################################

# replace recursive subset with R's default in-place semantics ====

# replace "AAB" using R's default in-place semantics:
lst_recin(
  lst, c("A", "A", "B"),
  rp = "THIS IS REPLACED WITH IN-PLACE SEMANTICS"
)
print(lst)



#############################################################################

# transform recursive subsets with R's default in-place semantics ====

lst_recin(lst, c("C", "A"), tf = \(x)x^2) # transforms lst$C$A

print(lst)


#############################################################################

# add/remove new recursive subsets with R's default in-place semantics ====

lst_recin(lst, c("C", "D"), rp = "NEW VALUE") # adds lst$C$D
print(lst)

lst_recin(lst, c("C", "A"), rp = NULL) # removes lst$C$A
print(lst) # notice lst$C$A is GONE


#############################################################################

# Modify View of List By Reference ====

x <- list(
 a = data.table::data.table(cola = 1:10, colb = letters[1:10]),
 b = data.table::data.table(cola = 11:20, colb = letters[11:20])
)
print(x)
myref <- lst_rec(x, "a")
address(myref) == address(x$a) # they are the same
tt_set(myref, col = "cola", tf = \(x)x^2)
print(x) # notice x has been changed


}
\concept{lst_}
