% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_setv.R
\name{ma_setv}
\alias{ma_setv}
\title{Find and Replace Present Values in mutable_atomic Objects}
\usage{
ma_setv(x, v, rp, invert = FALSE, NA.safety = TRUE)
}
\arguments{
\item{x}{a \link{mutable_atomic} \bold{variable}.}

\item{v}{non-missing (so no \code{NA} or \code{NaN}) atomic scalar to find.}

\item{rp}{atomic scalar giving the replacement value.}

\item{invert}{Boolean. \cr
If \code{FALSE} (default), the equivalent of \verb{x[which(x == v()] <- rp} is performed; \cr
If \code{TRUE}, the equivalent of \code{x[which(x != v)] <- rp} is performed instead.}

\item{NA.safety}{Boolean. \cr
just like in \link{which},
\code{NA} and \code{NaN} results in \code{x==v} should be ignored,
thus \code{NA.safety} is \code{TRUE} by default. \cr
However, if it is known that \code{x} contains no \code{NA}s or \code{NaN}s,
setting \code{NA.safety} to \code{FALSE} will increase performance a bit. \cr
NOTE: Setting \code{NA.safety = FALSE} when \code{x} does contain \code{NA}s or \code{NaN}s,
may result in unexpected behaviour.}
}
\value{
Returns: VOID. This function modifies the object by reference. \cr
Do not use assignment like \code{x <- ma_setv(x, ...)}. \cr
Since this function returns void, you'll just get \code{NULL}. \cr \cr
}
\description{
The \verb{ma_setv(x, v rp)} function performs the equivalent of \cr
\code{x[which(x == v)] <- rp} \cr
but using \link[=squarebrackets_PassByReference]{pass-by-reference semantics}. \cr
\cr
This is faster than using \code{sb_set(x, i = which(x == v), rp = rp)}. \cr
\cr
Inspired by \code{collapse::}\link[collapse]{setv},
but written in 'C++' through 'Rcpp'. \cr \cr
}
\examples{

x <- mutable_atomic(c(1:20, NA, NaN))
print(x)
ma_setv(x, 2, 100)
print(x)
}
