% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims}
\alias{ndims}
\title{Get Number of Dimensions}
\usage{
ndims(x)
}
\arguments{
\item{x}{the object to get the number of dimensions from.}
}
\value{
An integer, giving the number of dimensions \code{x} has. \cr
For vectors, gives \code{0L}. \cr \cr
}
\description{
\code{ndims(x)} is short-hand for \code{length(dim(x))}. \cr
}
\examples{

x <- 1:10
ndims(x)
obj <- array(1:64, c(4,4,3))
print(obj)
ndims(obj)


}
