% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{first}
\alias{last}
\alias{copy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:copy]{copy()}}, \code{\link[data.table:first]{first()}}, \code{\link[data.table:last]{last()}}}
}}

