% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_ins.R
\name{sb_before}
\alias{sb_before}
\alias{sb_after}
\alias{sb_before.default}
\alias{sb_after.default}
\alias{sb_before.array}
\alias{sb_after.array}
\alias{sb_before.factor}
\alias{sb_after.factor}
\alias{sb_before.list}
\alias{sb_after.list}
\alias{sb_before.data.frame}
\alias{sb_after.data.frame}
\title{Methods to Insert New Values Before or After an Index Along a Dimension}
\usage{
sb_before(x, ...)

sb_after(x, ...)

\method{sb_before}{default}(x, new, pos = 1, .attr = NULL, ...)

\method{sb_after}{default}(x, new, pos = length(x), .attr = NULL, ...)

\method{sb_before}{array}(x, new, margin, pos = 1, .attr = NULL, ...)

\method{sb_after}{array}(x, new, margin, pos = dim(x)[margin], .attr = NULL, ...)

\method{sb_before}{factor}(x, new, pos = 1, .attr = NULL, ...)

\method{sb_after}{factor}(x, new, pos = length(x), .attr = NULL, ...)

\method{sb_before}{list}(x, new, pos = 1, .attr = NULL, ...)

\method{sb_after}{list}(x, new, pos = length(x), .attr = NULL, ...)

\method{sb_before}{data.frame}(x, new, margin, pos = 1, .attr = NULL, ...)

\method{sb_after}{data.frame}(x, new, margin, pos = collapse::fdim(x)[margin], .attr = NULL, ...)
}
\arguments{
\item{x}{see \link{squarebrackets_immutable_classes} and \link{squarebrackets_mutable_classes}.}

\item{...}{further arguments passed to or from other methods.}

\item{new}{the new value(s). The type of object depends on \code{x}:
\itemize{
\item For atomic objects, \code{new} can be any atomic object.
However, if one wished the added values in \code{new} to be named,
ensure \code{new} is the same type of object as \code{x}.
For example: use matrix with column names for \code{new} when appending/inserting columns to matrix \code{x}.
\item For factors, \code{new} must be a factor.
\item For lists, \code{new} must be a (possible named) list.
\item For data.frame-like objects, \code{new} must be a data.frame.
}}

\item{pos}{a strictly positive single integer scalar (so no duplicates),
giving the position along the dimension (specified in \code{margin}),
before or after which the new values are added.}

\item{.attr}{a list,
giving additional potentially missing attributes to be added to the returned object. \cr
By default, concatenation strips attributes,
since the attributes of \code{x} and \code{new} may not be equal or even compatible. \cr
In the \code{attr} argument, the attributes of the merged object can be specified. \cr
Only attributes that are actually missing AFTER insertion will be added,
thus preventing overwriting existing attributes like names. \cr
One may, for example, specify \code{.attr = sb_a(x)} or \code{.attr = sb_a(new)}. \cr
If \code{NULL} (the default), no attributes will be added post-insert. \cr
If speed is important, \code{NULL} is the best option (but then attributes won't be preserved). \cr}

\item{margin}{a single scalar, giving the dimension along which to add new values.}
}
\value{
Returns a copy of the appended object.
}
\description{
The \code{sb_before()} method
inserts new values before some position along a dimension. \cr
The \code{sb_after()} method
inserts new values after some position along a dimension. \cr
\cr
These functions use a modified version of \code{abind::}\link[abind]{abind}
(see reference below).
}
\examples{

# atomic objects ====

x <- matrix(1:20 , ncol = 4)
print(x)
new <- -1 * x
sb_before(x, new, 1)
sb_before(x, new, 2)
sb_after(x, new, 1)
sb_after(x, new, 2)

#############################################################################


# factors ====

x <- factor(letters)
new <- factor("foo")
sb_before(x, new)
sb_after(x, new)


#############################################################################


# lists ====

x <- as.list(1:5)
new <- lapply(x, \(x)x*-1)
print(x)
sb_before(x, new)
sb_after(x, new)


#############################################################################


# data.frame-like objects ====

x <- data.frame(a = 1:10, b = letters[1:10], c = 11:20, d = factor(letters[1:10]))
new <- data.frame(e = 101:110)
sb_before(x, new, 2)
sb_after(x, new, 2)
new <- x[1,]
sb_before(x, new, 1)
sb_after(x, new, 1)

}
\references{
Plate T, Heiberger R (2016). \emph{abind: Combine Multidimensional Arrays}. R package version 1.4-5, \url{https://CRAN.R-project.org/package=abind}.
}
