% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mlen.R
\name{seq_mlen}
\alias{seq_mlen}
\title{Multiple seq_len}
\usage{
seq_mlen(x)
}
\arguments{
\item{x}{an integer vector giving the lengths.}
}
\value{
A list with \code{length(x)} elements, where each element is the result of \code{seq_len}.
}
\description{
Create multiple sequences of certain lengths at once. \cr
It is a vectorized version of \code{lapply(x, seq_len)}. \cr \cr
}
\examples{
seq_mlen(c(10, 10, 3))

}
