% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_names.R
\name{seq_names}
\alias{seq_names}
\title{Generate Integer Sequence From a Range of Names}
\usage{
seq_names(names, start, end, inv = FALSE)
}
\arguments{
\item{names}{a character vector of names. \cr
Duplicate names, empty names, or a character vector of length zero are not allowed.}

\item{start}{the name giving the starting index of the sequence}

\item{end}{the name giving the ending index of the sequence}

\item{inv}{logical, if TRUE,
the indices of all names EXCEPT the names of the specified sequence will be given.}
}
\value{
An integer vector.
}
\description{
Generate integer sequence from a range of names.
}
\examples{

x <- data.frame(a = 1:10, b = letters[1:10], c = factor(letters[1:10]), d = -1:-10)
ind <- seq_names(colnames(x), "b", "d")
sb_x(x, col = ind)

}
