% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\alias{ndim}
\alias{rdim}
\alias{s}
\alias{s.default}
\title{Get Size Properties of an Object}
\usage{
ndim(x)

rdim(x)

s(x, ...)

\method{s}{default}(x, m = if (is.null(dim(x))) 0L else dim(x), ...)
}
\arguments{
\item{x}{a vector, array, or data.frame.}

\item{...}{further arguments passed to or from methods.}

\item{m}{the margin.}
}
\value{
For \code{ndim()}: \cr
An integer, giving the number of dimensions \code{x} has. \cr
For vectors, gives \code{0L}. \cr
\cr
For \code{rdim()}: \cr
An integer vector, giving the range \code{1:ndim(x)}. \cr
For vectors, gives \code{0L}. \cr
\cr
For \code{s()}: \cr
A numeric vector giving the size(s) of the object at the given margin(s).
}
\description{
\code{ndim(x)} is short-hand for \code{length(dim(x))}. \cr
\cr
\code{s(x, m)} gets the size of an object \code{x} along margin \code{m}. \cr
If \code{m == 0}, \code{s()} gets the length of an object. \cr
If \code{m > 0}, \code{s()} gets \code{dim(x)[m]} of the object. \cr
\cr
}
\examples{
x <- 1:10
ndim(x)
obj <- array(1:64, c(4,4,3))
print(obj)
ndim(obj)
}
