% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcseq.R
\name{slcseq_x}
\alias{slcseq_x}
\alias{slcseq_x.default}
\alias{slcseq_rm}
\alias{slcseq_rm.default}
\alias{slcseq_set}
\alias{slcseq_set.default}
\title{Efficient Sequence-based Subset Methods on (Long) Vectors}
\usage{
slcseq_x(x, ...)

\method{slcseq_x}{default}(x, from = NULL, to = NULL, by = 1L, ..., use.names = TRUE)

slcseq_rm(x, ...)

\method{slcseq_rm}{default}(x, from = NULL, to = NULL, by = 1L, ..., use.names = TRUE)

slcseq_set(x, ...)

\method{slcseq_set}{default}(x, from = NULL, to = NULL, by = 1L, inv = FALSE, ..., rp, tf)
}
\arguments{
\item{x}{an atomic object. \cr
For \code{slcseq_set} it must be a \link{mutable_atomic} \bold{variable}.}

\item{...}{see \link{squarebrackets_method_dispatch}}

\item{from, to, by}{see \link{cp_seq}.}

\item{use.names}{Boolean, indicating if flat names should be preserved. \cr
Note that, since \code{slcseq} operates on flat indices only, \code{dimnames} are always dropped.}

\item{inv}{Boolean, indicating whether to invert the sequence. \cr
If \code{TRUE},
\code{slcseq_set()} will apply replacement/transformation on all elements of the vector,
\bold{except} for the elements of the specified sequence.}

\item{rp, tf}{see \link{squarebrackets_modify}.}
}
\value{
Similar to the \code{sb_} methods.
}
\description{
The \code{slcseq_} - methods are similar to the \code{sb_} - methods,
except they don't require an indexing vector,
and are designed for memory efficiency. \cr \cr
}
\examples{


x <- mutable_atomic(1:1e7)

# extract:
slcseq_x(x, 1, 10)

# reverse:
slcseq_x(x, -1i, 1) |> head()

# remove:
slcseq_rm(x, 1, -11i) # all elements except the last 10

# replace every other element:
x <- mutable_atomic(1:1e7)
slcseq_set(x, 2, -1i, 2, rp = -1)
head(x)

# replace all elements except the first element:
x <- mutable_atomic(1:1e7)
slcseq_set(x, 1, 1, inv = TRUE, rp = -1)
head(x)

}
