% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride_eval.R
\name{stride_eval}
\alias{stride_eval}
\alias{eval_stride}
\alias{eval_stride.default}
\alias{is.stride}
\alias{formula2stride}
\title{Evaluate Stride Object}
\usage{
eval_stride(stride, x)

\method{eval_stride}{default}(stride, x)

is.stride(x)

formula2stride(form, x)
}
\arguments{
\item{stride}{see \link{squarebrackets_stride}.}

\item{x}{a (long) atomic vector.}

\item{form}{a formula, as described in \link{squarebrackets_stride}.}
}
\value{
\bold{Using \code{stride_v()}} \cr
The original stride object, but as a list. \cr
\cr
\cr
\bold{Using \code{stride_seq()} or \code{stride_ptr()}} \cr
A list with at least the following elements: \cr
\cr
\bold{\verb{$start}}: \cr
The actual starting point of the sequence. \cr
This is simply \code{from} translated to regular numeric. \cr
\cr
\bold{\verb{$end}}: \cr
The \bold{actual} ending point of the sequence. \cr
This is not the same as \code{to}. \cr
For example, the following code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{seq(from = 1L, to = 10L, by = 2L)
#> [1] 1 3 5 7 9
}\if{html}{\out{</div>}}

specifies \code{to = 10L}. \cr
But the sequence doesn't actually end at \code{10}; it ends at \code{9}. \cr
Therefore, \code{stride_seq(x, 1, 10, 2) |> eval_stride()} will return \code{end = 9}, not \code{end = 10}. \cr
This allows the user to easily predict where an sequence given in
\link{stride_seq}/\link{stride_ptrn} will actually end. \cr
\cr
\bold{\verb{$len}}: \cr
The actual vector length the sequence would be,
given the translated parameters. \cr \cr
}
\description{
\code{eval_stride()} evaluates a stride or formula object. \cr
This can be handy to check exactly what parameters will be fed to the \code{long_} methods. \cr
\cr
\code{formula2stride()} converts a formula to a stride_seq or stride_ptrn object.
}
\examples{

# extract all elements of x with the name "a":
nms <- c(letters, LETTERS, month.abb, month.name) |> rep_len(1e6)
x <- mutatomic(1:1e6, names = nms)
head(x)
stride <-  stride_v(names(x), v = "a")
long_x(x, stride) |> head()


# find all x smaller than or equal to 5, and replace with `-1000`:
stride <- stride_v(x, v = c(-Inf, 5))
long_set(x, stride, rp = -1000L)
head(x, n = 10)




x <- mutatomic(1:1e7)

# extract elements 2 to 9
long_x(x, ~ 2:9:1:1)

# reverse:
long_x(x, ~ 9:2:1:1)

# remove all elements except the last 10:
long_x(x, ~ 1:(.N - 10):1:-1)

# replace every other element:
x <- mutatomic(1:1e7)
long_set(x, ~ 2:.N:2:1, rp = -1)
head(x)

# replace all elements except the first element:
x <- mutatomic(1:1e7)
long_set(x, ~1:1:1:-1, rp = -1)
head(x)


# extract pattern c(TRUE, FALSE, FALSE, TRUE) from first 20 elements:
x <- mutatomic(1:1e7)
long_x(x, ~ 1:20:c(TRUE, FALSE, FALSE, TRUE):1)

}
