% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride_ptrn.R
\name{stride_ptrn}
\alias{stride_ptrn}
\alias{ptrn_len}
\title{Pattern Sequence-based stride}
\usage{
stride_ptrn(start, end, ptrn, use = 1L)

ptrn_len(start, end, ptrn)
}
\arguments{
\item{start}{natural scalar giving the starting point of the sequence.}

\item{end}{natural scalar giving the ending point of the sequence.}

\item{ptrn}{a logical vector, giving the pattern. \cr
\code{ptrn} must contain at least 1 \code{TRUE} and 1 \code{FALSE} element,
to avoid ambiguity. \cr
\code{ptrn} must have length that fulfils all of the following conditions:
\itemize{
\item it is \verb{>= 2}
\item it is \verb{<= length(start:end)}
\item it is \verb{<= 2^31 - 1}
}}

\item{use}{\code{1} to select indices \code{(start:end)[ptrn]}; \cr
\code{-1} to select all indices \bold{except} \code{(start:end)[ptrn]} \cr}
}
\value{
An object of class "stride".
}
\description{
\code{stride_ptrn()} specifies a patterned sequence,
in the form of \code{(start:end)[ptrn]},
for use in the \link[=long_x]{long_} methods. \cr
\cr
For example: \cr
the sub-set operation \code{long_x(x, stride_ptrn(start, end, ptrn))} \cr
is conceptually equivalent to \cr
\code{x[(start:end)[ptrn]]} \cr
\cr
\code{ptrn_len(start, end, ptrn)} calculates the length of \code{(start:end)[ptrn]},
without actually allocating said vector. \cr
\cr
}
\examples{


x <- mutatomic(1:1e7)

long_x(x, stride_ptrn(2, 20, c(TRUE, FALSE, FALSE, TRUE)))

long_x(x, stride_ptrn(20, 2, c(TRUE, FALSE, FALSE, TRUE)))

long_x(x, stride_ptrn(2, 20, c(TRUE, FALSE, FALSE, TRUE), -1)) |> head()

long_x(x, stride_ptrn(20, 2, c(TRUE, FALSE, FALSE, TRUE), -1)) |> head()

}
\seealso{
\link{squarebrackets_stride} \cr
}
