% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_stridev.R, R/stride_pv.R
\name{stride_pv}
\alias{stride_pv}
\alias{countv}
\title{Value-based stride}
\usage{
stride_pv(p, v = NULL, na = FALSE)

countv(p, v = NULL, na = FALSE, use = 1)
}
\arguments{
\item{p}{class-less atomic vector of properties,
with the same length as \code{x}, and ideally related to \code{x} \cr
For example, \code{p} may be the character vector \code{names(x)},
the raw vector \code{broadcast::checkNA(x, "raw")}, the classless (i.e. raw data) values of \code{x},
or even the raw data values of another long vector with the same length as \code{x}.}

\item{v}{a scalar or vector, depending on the type of \code{p},
indicating what values in \code{p} to look for. \cr
Details are given in the sections below. \cr}

\item{na}{\code{TRUE}, \code{FALSE}, or \code{NA}, indicating what to do with \code{NA}s/\code{NaN}s. \cr
If \code{na = TRUE}, \code{NA}s/\code{NaN}s are included in the sub-set operation (i.e. \code{NA}s/\code{NaN}s are extracted, removed, replaced, etc.). \cr
If \code{na = FALSE}, \code{NA}s/\code{NaN}s are excluded from the sub-set operation (i.e. \code{NA}s/\code{NaN}s are not extracted, not removed, not replaced, etc.). \cr
If \code{na = NA}, \code{v} is ignored, and \bold{only} \code{NA} values are searched for the sub-set operation. \cr
See also the additional sections below.}

\item{use}{\code{1} to check for specified condition, \code{-1} to check for the negated condition (i.e. \code{!condition}).}
}
\value{
An object of class "stride".
}
\description{
\code{stride_pv()} is used in the \link[=long_x]{long_} methods to specify
sub-set operations based on values in an atomic vector of properties. \cr
\cr
\code{stride_pv()} can be used in the
\link[=long_x]{long_} methods
to perform some rather complex sub-setting operations; \cr
but for a very basic understanding, consider the following illustration. \cr
In the simplest terms,
the sub-set operation \code{long_x(x, stride_pv(p, v, na))} \cr
is conceptually equivalent to the following: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
x[ifelse(is.na(p), na, p == v)] # if `na` is `TRUE` or `FALSE`
x[is.na(p)] # if `na` is `NA`

}\if{html}{\out{</div>}}

\code{countv()} is a helper function, that counts how often \code{v} appears in \code{p}.
The sections further below give more details. \cr
}
\section{The Basic Idea}{


The basic idea is as follows. \cr
Let \code{x} and \code{p} be 2 atomic vectors of the same length
(but they don't have to be of the same type). \cr
Let \code{v} be some atomic scalar of the same type as \code{p}. \cr
Given the result of the condition \code{y == v},
the basic idea is to perform the following sub-set operations: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
long_x(x, stride_pv(p, v))            # ==> x[p == v]
long_set(x, stride_pv(p, v), rp = rp) # ==> x[p == v] <- rp
long_set(x, stride_pv(p, v), tf = tf) # ==> x[p == v] <- tf(x[p == v]) 

}\if{html}{\out{</div>}}

The above is with the default argument specification \code{use = 1}. \cr
Of course one can invert the relationship by specifying argument \code{use = -1},
to get something like the following:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
long_x(x, stride_pv(p, v), use = -1)             # ==> x[p != v]                  
long_set(x, stride_pv(p, v), use = -1, rp = rp)  # ==> x[p != v] <- rp
long_set(x, stride_pv(p, v), use = -1, tf = tf)  # ==> x[p != v] <- tf(x[p != v])

}\if{html}{\out{</div>}}

And \code{p} is allowed to be the same vector as \code{x}, of course. \cr
\cr
This basic idea, however, can become more complicated,
depending on the atomic type of \code{p}, which is discussed in the next section. \cr
\cr
\cr
}

\section{Details per Atomic Type}{

\bold{Logical, Raw, Complex} \cr
For \code{p} of type \code{logical}, \code{raw}, and \code{complex},
\link{stride_pv} works exactly as explained in the previous section. \cr
\code{p} and \code{v} must be of the same atomic type. \cr
\cr
\cr

\bold{Numeric} \cr
For \code{p} of type \code{integer} or \code{double} (collectively referred to as "numeric"),
the basic idea laid-out before still holds: \cr
one can use atomic vector \code{p} and atomic scalar \code{v} to perform sub-set operations like \cr
\code{x[p == v]}. \cr
\cr
But one may be more interested in a range of numbers, rather than one specific number
(especially considering things like measurement error, and machine precision,
and greater-than/larger-than relationships). \cr
So for numeric \code{p}, one can also supply \code{v} of length \bold{2}. \cr
When \code{length(v) == 2L}, \code{long_} will check whether \code{p} is inside
(or outside if \code{use = -1}) the bounded range given by \code{v}. \cr
I.e. :

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
p >= v[1] & p <= v[2]

}\if{html}{\out{</div>}}

Note that \code{p} and \code{v} must both be numeric here,
but they don't have to be the same type. \cr
I.e. one can have \code{p} of type \code{integer} and \code{v} of type \code{double},
without problems. \cr
\cr
\cr

\bold{Character} \cr
For \code{p} of type \code{character},
the basic idea is still to do something like \code{x[p == v]}. \cr
\cr
When searching for string \code{v} for sub-setting purposes,
one may want to take into consideration things like different spelling,
spacing, or even encodings of the same string. \cr
Implementing every form of fuzzy matching or encoding matching is computationally intensive,
and also quite beyond the scope of this package. \cr
Instead, the user may supply a character vector \code{v} of arbitrary length,
containing all the variations
(in terms of spelling, spacing, encoding, or whatever)
of all the strings to look for. \cr
\cr
So if a vector is given for \code{v} (instead of a single string),
the following check is performed:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
p \%in\% v

}\if{html}{\out{</div>}}

\bold{NOTE} \cr
The order of \code{v} is \bold{irrelevant}. \cr
\cr
}

\section{Smaller Than, Greater Than}{

For numeric \code{p}, one can specify a range for \code{v}, as explained earlier. \cr
But note one can also specify something like \code{v = c(-Inf, 4)},
which essentially corresponds to the condition \code{y <= 4}. \cr
Thus, when \code{v} specifies a range, "greater-than" and "smaller-than" comparisons are also possible. \cr
\cr
\cr
}

\section{Handling NAs and NaN}{

We also have to handle the \code{NA}s and \code{NaN}s. \cr
The \code{na} argument can be used to specify what to do when
a \code{p} is \code{NA}. \cr
\cr
When \code{na = FALSE}, all \code{NA} values of \code{p} are always ignored. \cr
I.e. \code{long_x(x, stride_pv(p, v, na = FALSE), use = 1)} will not extract \code{NA}s/\code{NaN}s, \cr
and \code{long_x(x, stride_pv(p, v, na = FALSE), use = -1)} will not remove \code{NA}s/\code{NaN}s. \cr
\cr
When \code{na = TRUE}, \code{NA} values of \code{p} are always included. \cr
I.e. \code{long_x(x, stride_pv(p, v, na = TRUE), use = 1)} will also extract \code{NA}s/\code{NaN}s, \cr
and \code{long_x(x, stride_pv(p, v, na = TRUE), use = -1)} will also remove \code{NA}s/\code{NaN}s. \cr
\cr
One can also specify \code{na = NA},
which will ignore \code{v} completely,
and explicitly look for \code{NA}s/\code{NaN}s in \code{p} instead - like so:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
long_x(x, stride_pv(y, na = NA))                        # ==> x[is.na(y)]                     
long_x(x, stride_pv(y, na = NA), use = -1)             # ==> x[!is.na(y)]
long_set(x, stride_pv(y, na = NA), rp = rp)             # ==> x[is.na(y)] <- rp                 
long_set(x, stride_pv(y, na = NA), use = -1, rp = rp)  # ==> x[!is.na(y)] <- rp               
long_set(x, stride_pv(y, na = NA), tf = tf)             # ==> x[is.na(y)] <- tf(x[is.na(y)])    
long_set(x, stride_pv(y, na = NA), use = -1, tf = tf)  # ==> x[!is.na(y)] <- tf(x[!is.na(y)]) 

}\if{html}{\out{</div>}}

Handling \code{NA}s/\code{NaN}s works the same for all atomic types. \cr
For \code{p} of type \code{complex},
a value \code{p[i]} is considered \code{NA}
if \code{Re(p[i])} is \code{NA}/\code{NaN} and/or \code{Im(p[i])} is \code{NA}/\code{NaN}. \cr
\cr
Argument \code{v} is never allowed to contain \code{NA}/\code{NaN}. \cr
\cr
\cr
}

\section{All in One}{


Combining all of the above,
one can allocate indices in base 'R' to be equivalent to
the virtual indices produced by \verb{stride_pv(p, v, na), use)},
with the following code:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
# if `na = NA`:
ind <- which(is.na(p)) * sign(use)

# else if using scalar `v`:
ind <- which(ifelse(is.na(p), na, p == v)) * sign(use)

# else if using numeric range for `v`:
ind <- which(ifelse(is.na(p), na, p >= v[1] & p <= v[2])) * sign(use)

# else if using character vector for `v`:
ind <- which(ifelse(is.na(p), na, p \%in\% v)) * sign(use)

}\if{html}{\out{</div>}}
}

\examples{

# basic idea ====
nms <- c(letters, LETTERS, month.abb, month.name) |> rep_len(1e6)
x <- mutatomic(1:1e6, names = nms)
head(x)

# extract all elements of x with the name "a":
stride <-  stride_pv(names(x), v = "a")
long_x(x, stride) |> head()

# find all x smaller than or equal to 5, and replace with `-1000`:
stride <- stride_pv(x, v = c(-Inf, 5))
long_set(x, stride, rp = -1000L)
head(x, n = 10)


################################################################################
# Numeric range ====
#
x <- mutatomic(1:1e6)
head(x)
stride <- stride_pv(x, c(-Inf, 5))
long_x(x, stride) # x[x <= 5]


################################################################################
# Character ====
#
if(require(stringi)) {
  x <- stringi::stri_rand_shuffle(rep("hello", 1e5))
  head(x)
  stride <- stride_pv(x, "hello")
  long_x(x, stride) |> head() # find "hello"
  
  # find 2 possible misspellings of "hello":
  stride <- stride_pv(x, c("holle", "helol"))
  long_x(x, stride) |> head()
  
}




}
\seealso{
\link{squarebrackets_stride} \cr
}
