% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa01_squarebrackets_methods.R
\name{aaa01_squarebrackets_methods}
\alias{aaa01_squarebrackets_methods}
\alias{squarebrackets_methods}
\title{Methods Available in 'squarebrackets'}
\description{
This help page gives an overview of the methods available in 'squarebrackets'. \cr \cr
}
\section{Main Methods}{

The main methods of 'squarebrackets' use the naming convention \verb{<indexform>_<operation>}: \cr
\verb{<indexform>} tells you what form of indices the method uses; \cr
\verb{<operation>} tells you \bold{what operation} is performed. \cr
\cr
For the \verb{<indexform>} part, the following is available:
\itemize{
\item \code{ii_}: operates on subsets of atomic/recursive vectors/arrays
by \link[=squarebrackets_indx_fundamentals]{interior indices}.
\item \code{ss_}: operates on subsets of atomic/recursive arrays of any dimensionality
by \link[=squarebrackets_indx_fundamentals]{subscripts}.
\item \code{tt_}: operates on subsets of data.frames and atomic/recursive matrices
by \link[=squarebrackets_indx_fundamentals]{tabular indices}.
}

For the \verb{<operation>} part, the following is available:
\itemize{
\item \verb{_x}: extract, exchange, exclude, or duplicate (if applicable) subsets.
\item \verb{_mod}: modify subsets and return copy.
\item \verb{_set}: modify subsets using \link[=squarebrackets_PassByReference]{pass-by-reference semantics}.
\item \code{icom}: create indices to be used in \verb{[<-} to modify objects using the default Copy-On-Modify semantics. \cr \cr
}

To illustrate, let's take the methods used for extracting subsets
(\verb{_x}):
\itemize{
\item If \code{y} is a vector or array (of any dimension), \cr
\code{ii_x(y, i)} corresponds to \code{y[i]}.
\item If \code{y} is a 3d array, \cr
\code{ss_x(y, n(i, k), c(1, 3))} corresponds to \code{y[i, , k, drop = FALSE]}.
\item If \code{y} is a matrix or data.frame-like object, \cr
\code{tt_x(y, i, j)} corresponds to \code{y[i, j, drop = FALSE]}. \cr \cr
}
}

\section{Specialized Methods}{

The main methods of 'squarebrackets' are applicable for all supported types and classes
(provided the correct method for the correct dimensionality is used). \cr
'squarebrackets' also provides specialized methods specific to certain structures:
\itemize{
\item the \link[=lst_rec]{lst_} set of methods,
which deal with sub-set operations that are only relevant for (nested) lists,
but not for the other types of supported objects. \cr
\item the \link[=long]{long_} set of methods,
which deal with index-less sub-set operations,
that are only relevant for long atomic vectors. \cr \cr
}
}

\section{Other Methods}{

Besides the previously mentioned methods,
'squarebrackets' provides some additional methods that do not neatly fit into the above methods. \cr
These are the set of \code{sb_} methods,
which cover miscellaneous operations for atomic/recursive objects. \cr \cr
}

\section{Finding the Appropriate Help Pages}{


With knowledge of the naming convention of the main methods,
one can easily find out information about a particular method by usign the \verb{?} operator. \cr
So to find out about modifying objects by subscripts using Pass-by-Reference semantics,
type in: \cr
\code{?ss_set} \cr
\cr
}

