% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa02_squarebrackets_methods.R
\name{aaa02_squarebrackets_methods}
\alias{aaa02_squarebrackets_methods}
\alias{squarebrackets_methods}
\title{Methods}
\description{
This help page gives an overview of the methods available in 'squarebrackets'. \cr \cr
}
\section{Main Methods}{

The main methods of 'squarebrackets' use the naming convention \code{A_B}: \cr
\code{A} tells you on what kind of object and what kind of indices the method operates on; \cr
\code{B} tells you \bold{what operation} is performed. \cr
\cr
For the \code{A} part, the following is available:
\itemize{
\item \code{i_}: operates on subsets of \bold{atomic} objects by (flat/linear) indices.
\item \code{i2_}: operates on subsets of \bold{recursive} objects by (flat/linear) indices.
\item \code{ss_}: operates on subsets of \bold{atomic} objects by (dimensional) subscripts.
\item \code{ss2_}: operates on subsets of \bold{recursive} objects by (dimensional) subscripts.
\item \code{slice_}: uses \bold{index-less}, \bold{sequence-based}, and efficient operations on \code{mutatomic} objects.
\item \code{slicev_}: uses \bold{index-less}, \bold{value-based} and efficient operations on \code{mutatomic} objects. \cr \cr
}

For the \code{B} part, the following is available:
\itemize{
\item \verb{_x}: extract, exchange, or duplicate (if applicable) subsets.
\item \verb{_wo}: returns the original object \bold{without} the provided subsets.
\item \verb{_mod}: modify subsets and return copy.
\item \verb{_set}: modify subsets using \link[=squarebrackets_PassByReference]{pass-by-reference semantics}. \cr \cr
}

To illustrate, let's take the methods used for extracting subsets
(\verb{_x}). \cr
\cr
When \code{y} is atomic, the following holds (roughly speaking):
\itemize{
\item \code{i_x(y, i)} corresponds to \code{y[i]}
\item \code{ss_x(y, n(i, k), c(1, 3))} corresponds to \code{y[i, , k]} \cr \cr
}

When \code{y} is a list (i.e. recursive), the following holds (roughly speaking):
\itemize{
\item \code{i2_x(y, i)} corresponds to \code{y[i]} or \code{y[[i]]}
(depending on the arguments given in \code{i2_x()})
\item \code{ss2_x(y, n(i, k), c(1, 3))} corresponds to \code{y[i, , k]} or \code{y[[i, , k]]}
(depending on the arguments given in \code{ss2_x()}) \cr \cr
}
}

\section{Other Methods}{

Besides the main methods, 'squarebrackets' provides some additional methods that do not neatly fit into the above methods. \cr
\cr
First, there is the \link[=lst_rec]{lst_} set of methods,
which deal with sub-set operations that are only relevant for (nested) lists,
but not for the other types of supported objects. \cr
\cr
Second, there is \link{idx} method,
which works on both recursive and non-recursive objects,
and transforms/translates indices to be used in R's default copy-on-modify semantics. \cr
\cr
Finally, there are the \code{sb_} and \code{sb2_} sets of methods,
which cover miscellaneous operations for atomic and recursive objects,
respectively. \cr \cr
}

\section{Finding the Appropriate Help Pages}{


With knowledge of the naming convention of the main methods,
one can easily find out information about a particular method by usign the \verb{?} operator. \cr
So to find out about modifying recursive objects by subscripts using Pass-by-Reference semantics,
type in: \cr
\code{?ss2_set} \cr
\cr
}

