% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa05_squarebrackets_modify.R
\name{aaa05_squarebrackets_modify}
\alias{aaa05_squarebrackets_modify}
\alias{squarebrackets_modify}
\title{Regarding Modification}
\description{
This help page describes the main modification semantics
available in 'squarebrackets'. \cr \cr
}
\section{Base R's Native Modification Semantics}{

For most average users, R's default copy-on-modify semantics are fine. \cr
The \verb{_mod} methods use native semantics for atomic objects,
and shallow copy semantics for recursive objects. \cr
\cr
}

\section{Pass-by-Reference}{

'squarebrackets' provides
the \verb{_set} methods
to modify by reference,
meaning no copy is made at all. \cr
Pass-by-Reference is fastest and the most memory efficient. \cr
But it is also more involved than the other modification forms,
and requires more thought. \cr
See \link{squarebrackets_PassByReference} for more information. \cr
\cr
}

\section{Arguments \code{rp} and \code{tf} for Atomic Objects}{

\bold{Argument \code{rp}} \cr
The \code{rp} argument is used to replace the values at the specified indices
with the values specified in \code{rp}.
Using the \code{rp} argument in the modification methods,
corresponds to something like the following: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
x[...] <- rp

}\if{html}{\out{</div>}}

\bold{Argument \code{tf}} \cr
The \code{tf} argument is used to transform the values at the specified indices
through transformation function \code{tf}.
Using the \code{tf} argument
corresponds to something like the following: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
x[...] <- tf(x[...])

}\if{html}{\out{</div>}}

where \code{tf} is a function that \bold{returns} an object of appropriate type and size
(so \code{tf} should not be a pass-by-reference function). \cr
\cr
}

\section{Arguments \code{rp} and \code{tf} for Recursive Objects}{

The \code{rp} and \code{tf} arguments work mostly in the same way for recursive objects. \cr
But there are some slight differences. \cr
\cr
\bold{Argument \code{rp}} \cr
'squarebrackets' demands that \code{rp} is always provided as a list
in the S3 methods for recursive vectors, matrices, and arrays (i.e. lists). \cr
This is to prevent ambiguity
with respect to how the replacement is recycled or distributed over the specified indices \cr
(See \verb{Footnote 1} below). \cr
\cr
\bold{Argument \code{tf}} \cr
Most functions in (base) 'R' are vectorized for atomic objects, but not for lists \cr
(see \verb{Footnote 2} below). \cr
'squarebrackets' will therefore apply transformation function \code{tf} via \code{lapply},
like so: \cr

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
x[...] <- lapply(x[...], tf)

}\if{html}{\out{</div>}}
}

\section{Arguments \code{rp} and \code{tf} for data.frame-like Objects}{

Replacement and transformations
in data.frame-like objects are a bit more flexible than in Lists. \cr
\cr
\bold{Argument \code{rp}} \cr
\code{rp} is not always demanded to be a list for data.frame-like objects,
only when appropriate
(for example, when replacing multiple columns, or when the column itself is a list.) \cr
\cr
\bold{Argument \code{tf}} \cr
Every column in a data.frame is like an element in a list. \cr
Therefore, when transforming parts of a data.frame with the \code{tf} argument,
lapply` is used for transformations across multiple columns. \cr
\cr
\cr
}

\section{Recycling and Coercion}{

Recycling is not allowed in the modification methods. \cr
So, for example, \code{length(rp)} must be equal to the length of the selected subset,
or equal to \code{1}. \cr
\cr
When using Pass-by-Reference semantics,
the user should be extra mindful of the auto-coercion rules. \cr
See \link{squarebrackets_coercion} for details. \cr
\cr
}

\section{Footnotes}{

\bold{Footnote 1} \cr
Consider the following replacement in base 'R':

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <-list(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
x[1:2] <- 2:1

}\if{html}{\out{</div>}}

What will happen? \cr
Will the \code{x[1]} be \code{list(1:2)} and \code{x[2]} also be \code{list(1:2)}? \cr
Or will \code{x[1]} be \code{list(2)} and \code{x[2]} be \code{list(1)}? \cr
It turns out the latter will happen; but this is somewhat ambiguous from the code. \cr
To prevent such ambiguity in your code,
'squarebrackets' demands that \code{rp} is always provided as a list. \cr
\cr
\bold{Footnote 2} \cr
Most functions in (base) 'R' are vectorized for atomic objects, but not for lists. \cr
One of the reasons is the following: \cr
In an atomic vector \code{x} of some type \code{t},
every single element of \code{x} is a scalar of type \code{t}. \cr
However, every element of some list \code{x} can be virtually anything: \cr
an atomic object, another list,
an unevaluated expression, even dark magic like \code{quote(expr =)}. \cr
It is difficult to make a vectorized function for an object with so many unknowns. \cr
Therefore, in the vast majority of the cases,
one needs to loop through the list elements. \cr
\cr
}

\concept{modify}
