% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa07_squarebrackets_method_dispatch.R
\name{aaa07_squarebrackets_method_dispatch}
\alias{aaa07_squarebrackets_method_dispatch}
\alias{squarebrackets_method_dispatch}
\title{Method Dispatch of 'squarebrackets'}
\description{
This help page gives some additional details regarding the S3 method dispatch
used in 'squarebrackets'. \cr \cr
}
\section{Ellipsis}{

Due to how the S3 method dispatch system works in 'R',
all generic methods have the ellipsis argument (\code{...}). \cr
For the user's safety,
'squarebrackets' does check that the user doesn't accidentally
add arguments that make no sense for that method
(like specifying the \code{inv} argument when calling \link{ii_x}). \cr
\cr
\cr
}

