% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa11_squarebrackets_keywords.R
\name{aaa11_squarebrackets_keywords}
\alias{aaa11_squarebrackets_keywords}
\alias{squarebrackets_keywords}
\alias{keywords}
\title{Advanced Index Specification with Keywords}
\description{
Instead of the usual indexing types,
formulas can be used to specify indices in vectors, arrays, or data.frames that can evaluate keywords. \cr
\cr
The following keywords are available:
\itemize{
\item \code{.M}: the given margin/dimension; \code{0} if not relevant.
\item \code{.Nms}: the (dim)names at the given margin.
\item \code{.N}: the size of a given dimension (if \code{.M} is not \code{0}) or else the length of \code{x}.
\item \code{.I}: equal to \code{seq_len(.N)}.
\item \code{.bi(...)}: a \bold{function} to specify bilateral indices. See the \code{Details} section below.
\item \code{.ptrn(ptrn, start, end)}: a \bold{function} specifying a pattern of indices. \cr
It is defined as \code{(start:end)[ptrn]},
where \code{start} and \code{end} are numeric scalars and \code{ptrn} is a logical vector. \cr
By default, \code{start = 1L} and \code{end = .N}.
}

These keywords can be used inside formulas to specify more advanced indices. \cr

Here are a few examples of advanced indexing in arrays:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
x <- matrix(1:20, 5, 4)
dimnames(x) <- list(month.abb[1:5], month.abb[1:4])
print(x)
#>     Jan Feb Mar Apr
#> Jan   1   6  11  16
#> Feb   2   7  12  17
#> Mar   3   8  13  18
#> Apr   4   9  14  19
#> May   5  10  15  20


# select first half of rows, select first & last column:
tt_x(x, ~ 1:round(.N/2), ~ .bi(1, -1))
#>     Jan Apr
#> Jan   1  16
#> Feb   2  17


# extract where rownames contain "r" & where colnames DO NOT contain "r":
ss_x(x, ~ grep("r", .Nms), use = c(1, -2)) 
#>     Jan Feb
#> Mar   3   8
#> Apr   4   9


# reverse order of all dimensions:
ss_x(x, ~ .bi(-.I))
#>     Apr Mar Feb Jan
#> May  20  15  10   5
#> Apr  19  14   9   4
#> Mar  18  13   8   3
#> Feb  17  12   7   2
#> Jan  16  11   6   1


# select every other element for each dimension:
ss_x(x, ~ .ptrn(c(TRUE, FALSE)))
#>     Jan Mar
#> Jan   1  11
#> Mar   3  13
#> May   5  15


# select different range from each dimension:
x <- array(1:1000, 5:3)
start <- c(3, 1, 2)
ss_x(x, ~ (start[.M]):.N)
#> , , 1
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   23   28   33   38
#> [2,]   24   29   34   39
#> [3,]   25   30   35   40
#> 
#> , , 2
#> 
#>      [,1] [,2] [,3] [,4]
#> [1,]   43   48   53   58
#> [2,]   44   49   54   59
#> [3,]   45   50   55   60
}\if{html}{\out{</div>}}
}
\details{
\bold{Bilateral Indices} \cr
The \code{.bi(...)} function is used to specify bilateral indices. \cr
One can specify both positive and negative numbers. \cr
Positive numbers (for example \code{~ .bi(1:10)}) specify indices work like normally. \cr
Negative numbers specify indices from the end; \cr
i.e. \code{~ .bi(-2)} would be equivalent to \code{length(x) - 1} if using an \code{ii_} method,
or \code{dim(x)[.M] - 1} if using a \code{ss_} or \code{tt_} method. \cr
\cr
The usage of \code{.bi(...)} is similar to the \code{c(...)} function,
in that multiple vectors can be concatenated to one. \cr
And one can specify both negative and positive numbers together. \cr
For example: \code{~ .bi(-10:-1, 10:1)}. \cr
\cr
One can also use the other keywords inside \code{.bi()}. \cr
For example, \code{~ .bi(-.I)} can be used to reverse a vector,
or reverse a dimension of an array or data.frame. \cr
\cr
}
\concept{keywords}
