% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_by.R
\name{idx_by}
\alias{idx_by}
\title{Compute Grouped Indices}
\usage{
idx_by(f, r, grp, parallel = FALSE, mc.cores = 1L)
}
\arguments{
\item{f}{a subset function to be applied per group on \code{r}. \cr
The function must accept a character or integer vector as input,
and produce a character or integer vector as output. \cr
For example, to subset the last element per group, specify: \cr
\code{f = last}}

\item{r}{an integer or character vector,
giving the complete range of indices of an object. \cr
For example: \code{colnames(x)}, \code{1:nrow(x)}, etc.}

\item{grp}{a factor giving the groups.
Make sure its order corresponds to \code{i} and \code{r},
otherwise it makes no sense.}

\item{parallel, mc.cores}{see \link[collapse]{BY}.}
}
\value{
A vector of indices of the same type as \code{r}.
}
\description{
Given:
\itemize{
\item a sub-set function \code{f};
\item the complete range of indices \code{r} of some object \code{x};
\item and a grouping factor \code{grp};
}

the \code{idx_by()} function takes indices \code{f(r)} \bold{per group} \code{grp}. \cr
The result of \code{idx_by()} can be supplied to the indexing arguments
(see \link{squarebrackets_indx_args})
of: \cr
\link{sb_x}, \link{sb_rm}, \link{sb_mod}, \link{sb_set}, or \link{sb_coe}, \cr
to perform \bold{grouped} subset operations. \cr
\cr
}
\examples{


# vectors ====
(a <- 1:20)
(grp <- factor(rep(letters[1:5], each = 4)))

# get the last element of `a` for each group in `grp`:
i <- idx_by(last, seq_along(a), grp)
sb_x(cbind(a, grp), row = i)


# data.frame ====
x <- data.frame(
  a = sample(1:20),
  b = letters[1:20],
  group = factor(rep(letters[1:5], each = 4))
)
print(x)
# get the first row for each group in data.frame `x`:
row <- idx_by(first, 1:nrow(x), x$group)
sb_x(x, row)
# get the first row for each group for which a > 10:
x2 <- sb_x(x, filter = ~ a > 10)
row <- na.omit(idx_by(first, 1:nrow(x2), x2$group))
sb_x(x2, row)
 

}
