% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx_r.R
\name{idx_r}
\alias{idx_r}
\title{Compute Integer Index Range}
\usage{
idx_r(x, m = 0L, start = NULL, end = NULL, by = 1L)
}
\arguments{
\item{x}{the object to index.}

\item{m}{integer or complex number, giving the margin. \cr
For non-dimensional objects or for flat indices, specify \code{m = 0L}.}

\item{start}{integer or complex, of the same length as \code{m} or of length 1,
specifying the start point.}

\item{end}{integer or complex, , of the same length as \code{m} or of length 1,
specifying the end point.}

\item{by}{integer, of the same length as \code{m} or of length 1,
specifying the step size. \cr}
}
\value{
If \code{length(m) == 1L}: a vector of numeric indices. \cr
\cr
If \code{length(m) > 1L}: a list of the same length as \code{m},
containing numeric vectors of indices. \cr
}
\description{
\code{idx_r()} computes integer index range(s). \cr \cr
}
\details{
\bold{start, end, by} \cr
If \verb{start, end} are not specified,
using \code{by} will construct the following sequence: \cr
If \code{by} is positive, \code{seq.int(1L, n, by)}. \cr
If \code{by} is negative, \code{seq.int(n, 1L, by)}. \cr
Where \code{n} is the maximum index
(i.e. \code{length(x)} or \code{dim(x)[m]}, depending on the situation). \cr
\cr
If \verb{start, end, by} are all specified,
\code{by} is stored as \code{abs(by)},
and the sign  of \code{by} is automatically adjusted to ensure a sensible sequence is created. \cr
\cr
So, for example,
to specify something like \code{n:1}, one can simply call \code{idx_r(x, m, by = -1L)}. \cr \cr

\bold{Heterogeneous Complex input} \cr
The user can specify complex vectors with varying imaginary parts for
\code{start} and \code{end}. \cr
Except for this difference,
the complex vectors are translated to integer vectors as explained in
\link{squarebrackets_indx_args}. \cr \cr

\bold{Multiple indices at once} \cr
The \code{idx_r()} function can compute indices for multiple dimensions at once,
by specifying a vector for \code{m}. \cr \cr
}
\examples{

x <- data.frame(
  a = 1:10, b = letters[1:10], c = factor(letters[1:10]), d = -1:-10
)
print(x)
ind1 <- idx_r(x, 1, 2, 2-1i) # rows 2:(nrow(x)-1)
sb2_x(x, ind1) # extract the row range

x <- array(1:125, c(5,5,5))
dims <- 1:3
sub <- idx_r(x, dims, 2, 2-1i) # 2:(n-1) for every dimension
sb_x(x, sub, dims) # same as x[ 2:4, 2:4, 2:4, drop = FALSE]

x <- letters
x[idx_r(x, 0, 2, 2-1i)]

}
