% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{first}
\alias{last}
\alias{copy}
\alias{address}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{address}}, \code{\link[data.table]{copy}}, \code{\link[data.table:last]{first}}, \code{\link[data.table]{last}}}
}}

