% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{s}
\alias{s}
\alias{s.default}
\title{Get the Size of an Object Along a Margin}
\usage{
s(x, m = 0L, ...)

\method{s}{default}(x, m, ...)
}
\arguments{
\item{x}{an object}

\item{m}{the margin.}

\item{...}{further arguments passed to or from methods.}
}
\value{
A numeric scalar.
}
\description{
\code{s(x, m)} gets the size of an object \code{x} along margin \code{m}. \cr
If \code{m == 0}, \code{s()} gets the length of an object. \cr
If \code{m > 0}, \code{s()} gets \code{dim(x)[m]} of the object. \cr
}
\examples{

x <- array(1:64, c(4,4,3))
print(x)
ss <- n(c(s(x, 1), 1), c(s(x, 3), 1))
ss_x(x, ss, c(1,3))



}
