% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride_seq.R
\name{stride_seq}
\alias{stride_seq}
\title{Regular Sequence-based stride}
\usage{
stride_seq(from, to, by)
}
\arguments{
\item{from}{natural scalar giving the starting point of the sequence.}

\item{to}{natural scalar giving the maximalle allowed end value.}

\item{by}{natural scalar giving the step-size.}
}
\value{
An object of class "stride".
}
\description{
\code{stride_seq()} specifies a regular sequence for use in the \code{long_} methods. \cr
For example: \cr
\code{long_x(x, stride_seq(from, to, by))} \cr
is conceptually equivalent to \cr
\code{x[seq(from, to, by)]} \cr
}
\examples{


x <- mutatomic(1:1e7)

# extract elements 2 to 9
long_x(x, stride_seq(2, 9, 1))

# reverse:
long_x(x, stride_seq(9, 2, 1))

# remove elements 1 to length(x) - 10:
long_x(x, stride_seq(1, length(x) - 10, 1), -1)

# replace every other element:
x <- mutatomic(1:1e7)
long_set(x, stride_seq(2, length(x), 2), rp = -1)
head(x)

# replace all elements except the first element:
x <- mutatomic(1:1e7)
long_set(x, stride_seq(1, 1, 1), use = -1, rp = -1)
head(x)

}
\seealso{
\link{squarebrackets_stride} \cr
}
